Change deepcopy to shallow dict copy.
The deep copy operation is very expensive; particle generation performance is
30% faster with this one change.

It's also unecessary because all values in self.contents are immutable, hence
it's impossible for a change to the copy to affect the original.

If there were a mutable object in self.contents, then the above may no longer
be true.